requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.3
djangorestframework==3.15.2
idna==3.10
mysql-connector-python==9.0.0
psycopg2-binary==2.9.10
pyTelegramBotAPI==4.23.0
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
telebot==0.0.5
tzlocal==5.2
urllib3==2.2.3
